feat(r): Add R Driver Manager (#365)
This is a proposal for an ADBC driver manager for R (#227). The scope of
this driver manager as compared to the Python version is limited: this
driver manager is about providing the minimum infrastructure required to
package and test drivers. It is designed to support but not replace
in-development user-facing packages (e.g.,
https://github.com/r-dbi/adbc).

This PR includes an 'adbcdrivermanager' package that includes two dummy
drivers that were required to enable full test coverage of the R
wrappers around adbc.h. The package also inclues `adbc_*()` functions
that mirror the `Adbc*()` functions in the C API pretty much 1:1 except
for the init/release cycles which were modified slightly to make it
harder for a user to crash an R session by accident.

The wrappers in the R package use R's "external pointer" to keep ADBC
objects alive and manage pointer dependencies (e.g., a statement created
from a connection will keep the connection object alive until the
statement is garbage collected). The external pointers carry a class of
`adbc_(database|connection|statement)` but allow for subclass so that
R-specific init methods can be defined (and perhaps other S3 methods in
a user-facing package). These external pointers also have the ability to
set key/value R-level metadata (helpful for stuff like print methods).

Drivers are slightly different because the pointer to them doesn't
really exist until the database in initialized (right?)...instead they
are an object with an init function or DLL/entrypoint (a driver factory
of sorts).