fix(r/adbcdrivermanager): Check that bootstrap.R exists before trying to run it (#437)
After trying the install instructions for the R driver manager added in
#365, we get rather scary message (although the install works just
fine):

```
* installing *source* package ‘adbcdrivermanager’ ...
** using staged installation
Fatal error: cannot open file 'bootstrap.R': No such file or directory
Fetched ADBC from https://github.com/apache/arrow-adbc
** libs
```

This happens because bootstrap.R is in `.Rbuildignore`, so it doesn't
end up in the built tarball. It's referenced from `configure` though
(for when somebody is installing from the source directory, not a built
tarball). This PR just adds a check for the file before attempting to
run it.