feat(r/adbdpostgresql): Package postgresql driver for R (#511)
This PR packages c/driver/postgresql for R on MacOS and Linux. I'd like
to defer the Windows issue to another PR because it's slightly more
complicated (unfortunately I think it will involve shipping a vendored
copy of libpq).

I added some minor changes to the driver so that the README
worked...bulk ingest was only supported for int64 and string, but
integer and double were needed for the existing README to work. The
example in the README segfaults after 700 ish rows anyway (I limit the
example to 100 rows) so I could back those changes out. I also get
intermittent segfaults in an interactive session that I wasn't able to
diagnose but I don't think those are a result of the R package setup
(they are probably related to deleting or inspecting connections in an
unexpected order).

---------

Co-authored-by: David Li <li.davidm96@gmail.com>