refactor(c/driver/postgresql): Factor out COPY reader and test it independently (#636)
I think there are some holes here but I think it's ready for a round of
review!

I'm aware that this may be slower than the current approach (notably:
virtual method calls and more assertions about buffer sizes). At this
point I think that type coverage and safety are worth it, and at least
this approach allows one to write benchmarks to determine if this is, in
fact, a bottleneck.

A few possible follow-ups:

- Add support for more types. Date/time/timestamp are highest on my
personal list.
- Do a similar refactoring for generating COPY data. In particular, bulk
inserts are very slow with the current approach and we could probably
speed up the bulk ingest considerably using COPY there, too.

I think the 374 exit code on Windows is not related to this PR since it
also occurs in the sqlite driver.