feat(r/adbcsnowflake): Package Snowflake driver for R (#638)
Snowflake for R!! Go in R!

The basic idea is to use `go build` to build static library and link it.
It's probably possible to do it using `c-shared` build mode but the
built package needs to be relocatable and `c-archive` seems to work.
This works on Mac (x86), Windows, and Linux. MacOS M1 doesn't work
because of a bug in the Go library that is resolved but not upstreamed
to the Snowflake Go connector yet. Windows requires R >= 4.2...it's
probably possible to build a binary distribution that will work on
earlier versions but it involves fiddling with PATH and is also a battle
for another day.

On install, the `go` binary is downloaded if it's not installed. This is
normally a no-no for CRAN packages but the CRAN battle is a question for
another day. For now, this enables somebody to follow along with the
readme and it should just work.

In addition to CI, it builds in CRAN's winbuilder (
https://win-builder.r-project.org/qc25F4zL5Bx2 ).

If you'd like to take it for a spin:

```r
# remotes::install_github("apache/arrow-adbc/r/adbcdrivermanager", build = FALSE)
# remotes::install_github("apache/arrow-adbc/r/adbcsnowflake#638", build = FALSE)
library(adbcdrivermanager)

# Use the driver manager to connect to a database. This example URI is
# <user>:<pass>@wt78143.<aws region>.aws/SNOWFLAKE_SAMPLE_DATA/TPCH_SF1?role=ACCOUNTADMIN
uri <- Sys.getenv("ADBC_SNOWFLAKE_TEST_URI")
db <- adbc_database_init(adbcsnowflake::adbcsnowflake(), uri = uri)
con <- adbc_connection_init(db)

stmt <- adbcdrivermanager::adbc_statement_init(con)
adbcdrivermanager::adbc_statement_set_sql_query(
  stmt,
  "SELECT * FROM REGION ORDER BY R_REGIONKEY"
)

stream <- nanoarrow::nanoarrow_allocate_array_stream()
adbcdrivermanager::adbc_statement_execute_query(stmt, stream)
tibble::as_tibble(stream)
```