chore(r): Clean up R verification tasks + ensure all drivers are verified (#659)
Ok! I checked the logs to make sure that all the drivers are actually
running and passing CMD check. Again, I'm punting on Windows. As part of
verification I run the bash verification script on Git bash on Windows
which works for the R portion...obviously it's better if it's in the
Powershell version but I don't have the bandwidth to make that happen
before 0.4.