feat(c/driver/postgresql): handle non-SELECT statements
Before we tried to infer the query schema before COPY by wrapping
it in a "SELECT * FROM (...) LIMIT 0".  This broke if the query
was (for example) a CREATE or UPDATE.  Instead, use a prepared
statement to infer instead.  If we find that there are no result
columns, then execute without the COPY path.

Also, test what happens with "INSERT INTO ... RETURNING" (this
works with COPY).

Fixes #701.