chore(r/adbcdrivermanager): Clean up driver manager to prepare for CRAN release (#756)
Basically a dry-run of the actual release checklist (#754). The goal
here is not "world's best documentation", just "enough documentation to
pass the CRAN incoming check. After the initial submission (which is
gated by a human), most releases will be automatic (particularly for a
package like this one, which doesn't have a large surface contact with
third-party dependencies or vendored code).