chore(c/vendor): Make __int128 safe-math header opt-in (#1016)
When trying to submit adbcpostgres, I get:

```
* checking whether package 'adbcpostgresql' can be installed ... WARNING
Found the following significant warnings:
  vendor/portable-snippets/safe-math.h:171:9: warning: ISO C++ does not support '__int128' for 'psnip_safe_int128_t' [-Wpedantic]
  vendor/portable-snippets/safe-math.h:172:18: warning: ISO C++ does not support '__int128' for 'psnip_safe_uint128_t' [-Wpedantic]
See 'd:/RCompile/CRANguest/R-devel/adbcpostgresql.Rcheck/00install.out' for details.
```

This is the workaround that I'm including in the CRAN packaging branch
for 0.6.0 but it would be nice to avoid dealing with it on subsequent
submissions. The only function we use from this header is the
overflow-safe int64 multiplication for timestamp support in the postgres
driver...given that we're always multiplying by a fixed 1000 in that
line, we could also probably hard-code the check + test the last
supported date?