feat(glib): add Vala VAPI for GADBC
Vala VAPI file is generated for current GADBC 1.0 API version.

It is disable by default, but recommended to be enable when
a development package is distributed.

This commit force a 1.0 API versioning, replacing the old one
where 0 were used for GIR API version (gir_version), so now
shows '1.0' instead of '0.0'.

Also examples are provided for C and Vala.