ci: fix snowflake integration tests (#1187)
We use quotes around the table name in the `DropTable` method but we
don't use it in `CreateSampleTable`. Without the quotes, the table is
created as upper case, but we attempt to drop the table with a forced
lower case name. As a result the table doesn't get dropped and we fail
on the table creation because the table already exists.

So if we properly put quotes around the table creation, this should fix
the snowflake CI integration tests.