feat(r/adbcdrivermanager): Implement missing function mappings (#1206)
These are the functions that were added in the 1.1.0 version
specification. The functions added are:

- `adbc_connection_cancel()`
- `adbc_connection_get_statistic_names()`
- `adbc_connection_get_statistics()`
- `adbc_statement_execute_schema()`
- `adbc_(database|connection|statement)_get_option(|bytes|int|double)()`
- `adbc_error_from_array_stream()`

Two more bits of functionality were implemented but not directly
exposed:

- Option setting. There are other option-related improvements I'd like
to implement and the current option setting helpers have the "convert to
string" behaviour baked in, so those aren't wired up yet.
- Error details. I try not to keep the actual error around very long in
R land, instead preferring to copy the contents and discard the error,
so the error detail is not directly exposed.

The vast majority of the changes were because of the option setters and
getters. Most of the other functions do not share signatures that might
be templatable, but the setters and getters tipped the scale and
option-related semantics were moved to
options.cc/options.R/test-options.R. There were also some new types that
needed converting (bool, int64, double), which tipped the scale on
separating out all of the tests that check that invalid input won't
cause a crash or cause the input to be misinterpreted.

To properly test these I need to implement some changes in the dummy
drivers, and I'd like to save that for a separate PR as this one is
getting rather unwiedly.

@nbenn I also discovered a typo: if you were having any issues with
transaction support, `adbc_connection_rollback()` was actually calling
`AdbcConnectionCommit()` (!!!! 😬 ).