ci(r): Refactor CI configuration (#1210)
This PR moves R workflows into their own file since they don't depend on
any other jobs. It also adds an r-extended job that checks some of the
more obscure things (e.g., older R versions, valgrind). The gist of the
CI setup is:

- r-basic.yml: Makes sure that drivers build + pass tests with test
databases on Linux. These run when C, R, or Go code is modified. This
catches changes to the drivers.
- r-standard.yml: Checks driver manager and driver packages on Mac,
Windows, and Linux. These run when working on the R packages because
that is the time that MacOS and Windows-specific regressions are likely
to occur (e.g., when working on the build systems).
- r-extended.yml: Checks the way CRAN would check (on oldrel, release,
and devel versions of R on Mac, Windows, and Linux); checks older R
versions still in the tidyverse support grid on Windows (because the
build system changes so frequently) and Linux (because Go drivers don't
build on older R versions on Windows); runs tests with valgrind.

Closes #1138.

I opened tickets to debug the valgrind errors reported by the drivers.
Those CI jobs run weekly (reflecting the time I have available to debug
them) so they won't add noise to ongoing development.