refactor(r): Improve testing for ADBC 1.1 features in R bindings (#1214)
This PR implements option setting/getting in the "void" driver and
implements tests for the full grid of set/get by
string/bytes/integer/double by database/connection/statement. The error
detail information was also not implemented in the dummy driver and so
couldn't be tested (so there is an implementation of that here).

Implementing a driver that actually did this was sufficient work that I
did some rather heavy abstraction to make it easier to write. That
abstraction is not unlike a "driver framework" except it is (1) not
complete, since the void driver only needs options methods and (2)
doesn't provide any result helpers (building streams, etc.). It might be
worth porting the driver base to be more general but for now I'd like to
keep it constrained to what I need to test in R.

Closes #1126.