chore(r/adbcsnowflake): Update adbcsnowflake such that it can be submitted to CRAN (#1220)
Closes #1067.

The mailing list thread where I asked about this (
https://stat.ethz.ch/pipermail/r-package-devel/2023q3/009329.html )
noted a few possible requirements:

- Make sure to not download dependencies. If you do need to download,
bundle the dependencies separately and download them at install time
(with checksum verification). Because our dependencies are >5MB zipped
and contain paths longer than 100 characters, we need to do this. I
wired this up so you can test it locally by providing a `file://`
URL...when submission times come that URL and shasum can be hardwired
into the package.
- Make sure to not download Go itself and make sure to check Go
locations that might not be on PATH. I removed the go download for
now...if it becomes too hard of a constraint to satisfy it could be
reinstated and removed just prior to submission. Luckily, installing Go
is pretty straightforward.

I also added some updates to make the tests run with my local Snowflake
account and consoliated the Windows and MacOS and Linux builds since the
all use the same logic.

This will probably need to be workshopped during submission and, after
successful submission, distilled into some common place and applied to
the adbcflightsql package.