test(csharp/test/Drivers/Snowflake): Correct catalog field name in README, generify expected table name in CanClientExecuteUpdateUsingExecuteReader test (#1240)
README currently lists the field name as "database", when it should be
"catalog". Also add a few small notes that the config file is JSON and
there's a template in resources to use.

CanClientExecuteUpdateUsingExecuteReader() currently has the table name
ADBC_ALLTYPES hardcoded, ignoring the table value set in the
configuration. It can be made to use the config table name instead.