test(csharp/test/Drivers/Snowflake): Correct FLOAT min/max tests in .NET 4.7.2 (#1312)
.NET 4.7.2 will round up Double.MaxValue / round down Double.MinValue
when using Double.ToString(), resulting in infinity / -infinity stored
in Snowflake when passed into the INSERT query statement. This rounding
does not occur in .NET 6.0.

This PR corrects that behavior and sends a properly accepted
MaxValue/MinValue string when using .NET 4.7.2.

Notable that when the tests can use prepared statements, issues like
this should cease to happen.