chore(r): Housekeeping for R packages (#1399)
A few housekeeping updates:

- Updated documentation to reflect implementation/installation status in
the documentation
- Updated NEWS entries up to the current version number
- Upstreamed updates requested by CRAN regarding the Snowflake driver
(syncing most with the flightsql driver as well). Some of this could be
factored out into a common 'how to link a go library from R' package,
but it's probably too early for that since there isn't quite consensus
on what is actually going to be required yet.