python: switch to static versioning for simplicity (#1419)
We tried to derive versions from Git. But since we release off of a
branch, this meant that all non-branch builds would have version 0.0.0.
This breaks things like Polars that rely on checking the installed
version.

Instead, since we already bump versions as part of the release process,
just hardcode the version and drop most of miniver.

Fixes #1363.