test(c/driver/sqlite): ensure that test suite passes with system-provided SQLite on MacOS (#1499)
I opened up the C drivers and ran the tests and got a failing test: when
the driver compiles against system SQLite on MacOS, extension loading is
not allowed (the symbols aren't even provided nor defined in the
header). I added a define to make sure the R package could build on
MacOS, but the test needs it too.

I also added come CMake to check this...I also tried
`check_sybmol_exists` but this seems to require setting
`CMAKE_REQUIRED_LIBRARIES`/includes which seems like it might have some
global impact that we don't want. Checking for the text in the header
does the trick for me (and is maybe less likely to result in
accidentally building the library without extension support).

The other failing test I saw before seems to have been cleared up by the
SQLite driver refactor!

The `.cache` in the `.gitignore` is because clangd seems to put a lot of
files there.