feat(rust): add integration tests and some improvements (#1883)
This should be the last PR for the initial Rust implementation.

It includes:
- An integration test for the driver manager against the official SQLite
driver
- An integration test for the dummy driver against itself (comparing
output when using the Rust API and when using its exported version
through the driver manager)
- New info codes and options which have been added recently
- An enum for identifying statistics

Note that I also have an integration test for the driver manager against
the official PostgreSQL driver but it needs a running PostgreSQL
instance to be executed, so it cannot be easily integrated into the CI.
This test proved useful during development as it allowed me to catch
some bugs. The existence of this test is the reason why there is a
common module. Maybe we can add it in a follow-up PR?