fix(docs): correct snowflake options for bulk ingest (#2004)
It was brought up in #1997 that the currently published options for
snowflake bulk ingestion are incorrect in the docs. This corrects them
to the proper values that are consistent with the values in the
implementation.

This also adds new constants to the python `StatementOptions` enum for
the snowflake driver for users to reference.