fix(ci): Pin r-lib actions as a workaround for latest action updates (#2051)
The R check actions are currently failing because an update to r-lib
actions resulted in some quarto actions being invoked, and these have
not yet been whitelisted for use in Apache repositories. We don't
actually need quarto for anything in the R package; however, the action
is written in such a way that even though the step is skipped, the
action is still loaded (and therefore its name is checked against the
whitelist).

Unrelatedly, the extended R check was failing because I forgot to copy a
change from #1965 into one of the workflows that only runs once a week.