docs(r): fix some pkgdown website linking issues (#2154)
- Fix base URL of each sites. Due to the base URLs, the site search does
not appear to be working.
For example, I searched "adbc" in
<https://arrow.apache.org/adbc/current/r/adbcdrivermanager/index.html>
returns
<https://arrow.apache.org/reference/adbc_driver_log.html?q=adbc#ref-usage>
(On a related note, the search feature seems to stop working when a new
version is released and the base URL changes; we can confirm that the
search does not work in previous versions of the `arrow` package.)
- It seems that pkgdown (downlit) uses the first value in the URL field
of the DESCRIPTION file to autolinking, and bringing the pkgdown site
first seems to be the expected use. (e.g. tidyverse/glue#300)