refactor(c/driver/postgresql): Use GetObjectsHelper from framework to build objects (#2189)
This PR migrates the PostgreSQL driver to use the same
`GetObjectsHelper` as the SQLite driver (the one provided by the
framework). This involved a few updates to the helper (notably: ensure
the filter is passed to the `LoadXXX()` methods), and an update to the
result helper to make it easier to store iteration state along a result
as a member of the helper.