feat(csharp/src/Drivers): Add Databricks driver (#2672)
After the initial Spark work, the Databricks-specific features were
becoming more entwined in the code. There have been requests for
additional Databricks-specific features such as CloudFetch, Entra
authentication, and LZ4 compression to name a few.

This PR moves the Databricks-specific code out of the Spark driver and
into its own driver. It leverages the Thrift and Spark libraries that
exist in the Apache driver but teases out the Databricks capabilities
from the Spark capabilities. It does so by exposing the internals of the
Apache driver so that the Databricks driver can leverage the same
underlying libraries while still being able to offer differentiating
features and raises the branding of the driver from under the
Apache/Spark umbrella to the same level as BigQuery, FlightSQL, and
Snowflake.

---------

Co-authored-by: David Coe <>
Co-authored-by: Bruce Irschick <bruce.irschick@improving.com>