chore(r): Update CI setup to silence warnings for Go-based drivers (#2708)
Closes #2707.

This is more likely to be a change in Adbc somewhere since it's showing
up in all versions. Did we add anything recently that logged anything to
stderr or called assert or abort?

In the event it can't be avoided (it often can't), we can update the
warning fail level for go-based drivers since we don't distribute them
on CRAN.