feat(go/adbc/driver/snowflake): New setting to set the maximum timestamp precision to microseconds (#2917)
Introduces a new setting to set the maximum timestamp precision to
Microsecond. Setting this value will convert the default Nanosecond
value to Microsecond to avoid the overflow that occurs when a date is
before the year 1678 or after 2262.

Provides a fix for https://github.com/apache/arrow-adbc/issues/2811 by
creating a workaround that can be set by the caller.

---------

Co-authored-by: David Coe <>