fix(csharp/src/Drivers/BigQuery): Prevent callers from attempting to use the public project ID to create query jobs (#2966)
In some situations, the publicProjectId gets passed as the actual
project ID and causes an error when we try to create a query job:
 ```
Google.GoogleApiException : The service bigquery has thrown an
exception. HttpStatusCode is Forbidden.
Access Denied: Project bigquery-public-data: User does not have
bigquery.jobs.create permission in
      project bigquery-public-data.
``` 
This PR adds a check for the public project ID and treats it as if we need to detect the project ID for the caller.

Co-authored-by: David Coe <>