feat(python/adbc_driver_manager): Update python driver_manager to load manifests (#3018)
Adding a new function `AdbcDriverManagerDatabaseSetLoadFlags` so that
bindings can set the load flags into the Driver Manager's internal
`TempDatabase` to control the functionality of when to look for and load
manifests.

This also changes the Python driver_manager to load manifests by default
but allows passing a `load_flags` option to manually pass in load flags
to control the behavior. Also adding a new `driver_example_manifest.py`
recipe to the driver_example docs.