feat(csharp/src/Drivers/Databricks): Remove redundant closeoperation (#3093)
The driver was breaking with DBR 11.3.

[This PR introduced logic
](https://github.com/apache/arrow-adbc/commit/cde9e7b525b37dfdbde27a93d57ac214ea0701bb#diff-ea22827c908183e8cae28e5fcc37cef9d0c4b5637474262cffe163b6b78db9f5R386)which
throws errors via HandleThriftResponse. Previously, we had been silently
ignoring these CloseOperation failures; these errors occurred because
the operations can be closed when returning via DirectResults, and we
weren't properly checking for DirectResults.CloseOperation status for
Metadata queries. With this change, we avoid making redundant
CloseOperation requests upon dispose.

1. Metadata Request to server 
2. Metadata Response with DirectResults and CloseOperation 
3. Statement.Dispose 
4. DirectResults includes CloseOperation 
5. Skip redundant CloseOperation


### Testing
Added a test that only passes with the changes in this PR in DBR 11.3