feat(go/adbc/sqldriver): handle timestamp/time.Time values for input (#3109)
fixes #3103

Adding new cases to `arrFromVal` to allow for handling `time.Time` and
`arrow.Timestamp` and `arrow.Time32`/`arrow.Time64` types. This only
works when the parameter schema is provided by the FlightSQL server side
when utilizing a prepared statement or otherwise.

If we don't have the parameter schema, then it will error as usual.