feat(rust/core): add function to load driver manifests (#3099)
Following the pattern of the C++ driver manager and the various bindings
to it, we also have to add support for the Rust driver manager to handle
and load drivers via manifest files too. Since the Rust driver manager
doesn't bind to the C++ one like everyone else, we need to add the
support in rust directly to locate, open, and parse driver manifests.

This PR adds a new function `ManagedDriver::find_load_from_name` along
with constants for the `LoadFlags` to control what paths are searched.
The basic logic mirrors the C++ implementation. Unit tests are also
included.

To control the tests, two new features are added:
`driver_manager_test_lib` which expects an env var
`ADBC_DRIVER_MANAGER_TEST_LIB` to be set, just like the C++ driver
manager tests, to the path for an adbc library to load. And
`driver_manager_test_manifest_user` which controls whether the test that
interacts with the user's config directory should be run or not.

---------

Co-authored-by: David Li <li.davidm96@gmail.com>
Co-authored-by: Bryce Mecum <petridish@gmail.com>
Co-authored-by: Matthijs Brobbel <m1brobbel@gmail.com>
Co-authored-by: eitsupi <50911393+eitsupi@users.noreply.github.com>