fix(go/adbc/drivermgr): properly vendor toml++ (#3138)
The `drivermgr` package referenced the c/vendor directory in the CGO
building which doesn't work as that isn't within the Go module itself
and wouldn't get included when the module is served as a dependency. We
need to properly vendor the toml++ header file in the Go source so that
it is delivered and packaged with the module properly.