docs: add docs for driver manifests (#3176)
With the driver manager implementations for C/C++, Go, Rust and Python
updated to utilize and leverage driver manifests, we should properly
document how manifests work and what the format is.

This change updates the docs to include a description of the driver
manifest file and the process used to search for a driver manifest or
driver given just a name. This includes the Load Flags and example code.

---------

Co-authored-by: Sutou Kouhei <kou@cozmixng.org>