Fix what I think is a mistake, + reword
Is this something I just don't know about or just a typo? Usually when I use LD_LIBRARY_PATH, I add absolute paths to a folder containing .so files and then I have to pass the basename of the file. Here, you say I'd pass 'adbc_driver' to load 'libadbc_driver.so'. Is that a typo?

I also reworded the paragraph a bit.