refactor(rust)!: not to mark some enums as `non_exhaustive` (#3245)
As discussed in #3197, some enums should not be marked as
`non_exhaustive` because they are unlikely to be expanded in the future
without breaking changes.

One of the exceptions is `AdbcVersion`, but on our side, if this is
extended we have to make sure to change adbc_driver_manager as well, so
we also have to run the unstable `non_exhaustive_omitted_patterns` lint
with nightly toolchain.