fix(c/driver_manager,rust/driver_manager): establish standard platform tuples (#3313)
Add to the docs an authoritative definition of the platform tuples we'll
use for the driver manager (for OS, Arch, env). Update the C and Rust
driver managers to use these authoritative platform tuple strings.

This should unify everything to a specific way of identifying platform
tuples with Driver manifests.

---------

Co-authored-by: Bryce Mecum <petridish@gmail.com>