docs: fix safari rendering in manifest_load.mmd diagram (#3391)
Fixes rendering of the manifest_load.mmd diagram under Safari by
re-generating the SVG using mmdc. The original SVG checked into git was
generated by hand using a web tool (mermaidchart.org)) which has a
rendering bug under Safari. The YAML frontmatter this PR removes is not
supported by the mmdc cli (it's specific to mermaidchart.org) so
removing it is also required for the makefile to run correctly anyway.

This PR is an overall improvement but has one downside: When using dark
mode, the connector lines between graph nodes is slightly less bright.
We might be able to fix this with custom CSS (via `mmdc --cssFile`).

Resolves #3292.