docs: rework driver manager references across docs (#3388)
This PR does a somewhat comprehensive rework of how we refer to the
driver manager and makes a few tweaks in various places in the docs.

It started with the Python page for the driver manager but also removes
redundant descriptions of the driver managers in other languages in
favor of a common definition in the glossary. It extends the Python
driver manager docs a bit and also adds "driver manifest" as a term in
the glossary.

Resolves #3363
Resolves #3385

---------

Co-authored-by: Ian Cook <ianmcook@gmail.com>
Co-authored-by: David Li <li.davidm96@gmail.com>