fix(python/adbc_driver_manager): handle empty params in executemany (#3332)
Note that params in executemany is a sequence of parameter bindings (not
parameters in and of themselves).

"Old" behavior of params:
None and `[]` are a no-op.

"Current" behavior:
None and `[]` are the same as executing without params.

"Fixed" behavior:
None is the same as executing without params.
`[]` is a no-op.

Closes #3319.