feat(csharp/src/Drivers/BigQuery): Add support for specifying a location (#3494)
In the current iteration, no DefaultLocation is specified for the
internal BigQueryClient. Since this is not specified, the BigQuery API
makes this default to "US" which is causing failures for some users.
This PR adds a new parameter to set the DefaultLocation for the
BigQueryClient, which then uses that same value for creating datasets
and other location-based activitities.

---------

Co-authored-by: David Coe <>