fix(csharp/src/Drivers/BigQuery): improve selective handling of cancellation exception (#3615)
After adding cancellation functionality to BigQuery statements, there is
a report that `task was cancelled` messages are now appearing.

This is likely due to the change where the retry code stops processing
when it gets a `OperationCancelledExecption`.

This change now only exits early if the cancellation was requested on
the passed cancellation token.