fix(c/driver_manager): look for correct init symbol name (#3681)
- Add sanity assertions
- Actually test this end-to-end
- Populate lib_path properly so the right init symbol name gets
calculated

IMO, the control flow now is very convoluted because it intermixes
searching and loading, which led to this regression since it's hard to
tell if the return value is actually fully initialized. I'm still not
sure all code paths are correct. I may try to factor searching/loading
out separately.

Closes #3680.